+++ source/parsing/Preprocessor.cpp
@@ -881,7 +881,7 @@ bool Preprocessor::applyMacroOps(span<Token const> tokens, SmallVector<Token>& d
 
     // TODO: audit trivia usage here, use of dest.back(), etc
 
-    for (uint32_t i = 0; i < tokens.size(); i++) {
+    for (ptrdiff_t i = 0; i < tokens.size(); i++) {
         Token newToken;
 
         // Once we see a `" token, we start collecting tokens into their own
@@ -1042,7 +1042,7 @@ bool Preprocessor::expandMacro(MacroDef macro, MacroExpansion& expansion,
     };
     SmallMap<string_view, ArgTokens, 8> argumentMap;
 
-    for (uint32_t i = 0; i < formalList.size(); i++) {
+    for (ptrdiff_t i = 0; i < formalList.size(); i++) {
         auto formal = formalList[i];
         auto name = formal->name.valueText();
         if (name.empty())
